Minor fixes to factory.cpp and playerboard.cpp
Fixed the part where the board was getting redefined in the playerBoard constructor.  Fixed the conditional statements for row and column index check, where it would allow an index from 0 to max inclusive instead of 0 to max - 1 inclusive. Added conditional statements to the get(index) function from factory.cpp to only return the pointer if the index is within appropriate bounds.